feat(frontend): made all stream based routes child routes; fixed issues re multiple fetching
